catsVdogsAPI/frontend/requirements.txt
# frontend/requirements.txt
streamlit
pillow